A couple UDP fixes.
1. The Datagram class was broken. It was setting the addr on the class
not on instances! This was not a problem for the resolver where the
clients were all talking to the same server, but definitely showed up
with a server that was servicing multiple clients.

2. Packets in quick succession handled by a UDPSocket weren't working.
This is because it was doing no buffering between calls to receive().
Now any datagrams received over the socket while a receive() call is not
waiting will buffer.