Fair Queue scheduling and a bug fix.
sleep() is now used to force a context switch when an "early value" is
recevied from a Waiter or from a direct call to Queue.get. This doesn't
measurably slow down the threadbench Queue benchmark and it makes for
even scheduling across Queue consumers that are doing roughly the same
amount of work.

Also fixed a bug in wait() when an EarlyValue was returned.