DieselZMQService: Some internal and API changes.
Subclasses can now override convert_raw_data_to_message to allow for
using data inside of the received packet to identify the remote client.

There is also a new Message class that encapsulates the data to be sent,
the client identity it should map to and the ZeroMQ return routing
identity.

It's unfortunately a bit more complexity but it models some real-world
needs.