WebSockets API bugfix.
Previously we were passing a Request proxy object to the WebSocket
handler function. This caused a problem because that function is run in
a different green thread and thus a proxy for it doesn't exist on the
request context stack.

Passing a copy of the actual object solves that problem. It uses the
_get_current_object() API that is provided by Werkzeug. Though that
looks like a private API (leading underscore) it is referenced in the
Flask documentation as a way to get things done in certain cases.