A test for verifying correct ordering of events.
The actual bug that this tests for has been worked around at this point,
but if you want to try it yourself, uncomment the `diesel.sleep()` call
on line 419 of diesel/core.py. Run this test and it will fail.

A true fix and not just a workaround is forthcoming.