Handle non-RFC WebSockets behind proxies.
Proxies that are smart and inspect HTTP requests might change the
requests slightly by parsing the request, sending it on and waiting for
a resopnse. For non-RFC WebSockets, diesel was expecting the third key
to be included with the original request bytes. In the case of a proxy
like the one described above, that would not hold and the key would be
missing.

Also bumping the version to 3.0.7.