Fleshed out support for child loop keep-alive
It sort of worked before, but after the first death of a child loop, it
would become detached from its parent. Also, if a parent loop died, its
children with keep-alive set would live on.

Now child loops always retain their parents, and will stay alive as long
as the parent loop is alive. However, if their parent loop dies, they
will die along with it.

There is also a new `deaths` attribute on `Loop` instances to track the
number of times that a `Loop` has died. It proved useful for testing and
it will likely be useful for debugging as well.