A test for `DieselZMQService` durability
It currently fails. This is because the `_receive_incoming_messages`
child loop can die and doesn't come back. When it dies, the service
continues to run but in a zombie-like undead state where it no longer
received incoming messages.

This change also extends `DieselZMQService` to make it more testable.
The changes include an overridable `should_run` flag and the addition of
a reference to the `_receive_incoming_messages` `Loop` on a
`DieselZMQService` instance