Don't lazily remove timers
This was biting us because timers that were set to fire far in the
future (many seconds at least) were piling up in the .timers deque in
the hub even while they were being deleted from the .timer_map. This was
causing the hub to spend too much time clearing them out and it was
falling behind on actual timers it was supposed to service and not being
able to recover.

This change moves the canonical storage of Timers in .timers to a list.
This is better for deleting cancelled timers (.remove() on a list is
twice as fast as .remove on a deque). The .timers list is kept in
reverse sorted order so that Timers that are due can be popped off the
end.