Fix ConnectionPool pool_max bug
Normal ConnectionPools are able to be created at the module level,
outside of any diesel Loop. This is a useful feature. However, if you
specified a pool_max value when creating your ConnectionPool outside of
a Loop, you'd get an exception.

This fix changes the way the sentinel values are added to the
remaining_conns Queue in the ConnectionPool constructor. It adds them
without firing any diesel events.